+++ src/index.test.ts
@@ -44,13 +44,11 @@ test('creates a recording with default filename', async () => {
   const axiosInstance = axios.create()
   useAxiosRecordReplayAdapter({axiosInstance})
 
-  const {data} = await axiosInstance.get(
-    'https://jsonplaceholder.typicode.com/todos/1',
-  )
+  const {data} = await axiosInstance.get('https://reqres.in/api/users/7')
 
   const [recording] = fs.readdirSync('./recordings')
   expect(recording).toMatchInlineSnapshot(
-    `"3439973d2c1ff6cc118d7af4cf797551.json"`,
+    `"1ab499d565acf87f67ac738631d42d07.json"`,
   )
 
   const recordingContents = fs.readFileSync(`./recordings/${recording}`, 'utf8')
@@ -60,16 +58,19 @@ test('creates a recording with default filename', async () => {
     "{
       'request': {
         'method': 'get',
-        'path': '/todos/1'
+        'path': '/api/users/7'
       },
       'response': {
         'status': 200,
         'statusText': 'OK',
         'data': {
-          'userId': 1,
-          'id': 1,
-          'title': 'delectus aut autem',
-          'completed': false
+          'data': {
+            'id': 7,
+            'email': 'michael.lawson@reqres.in',
+            'first_name': 'Michael',
+            'last_name': 'Lawson',
+            'avatar': 'https://s3.amazonaws.com/uifaces/faces/twitter/follettkyle/128.jpg'
+          }
         }
       }
     }"
@@ -86,25 +87,25 @@ test('creates a recording with filename prefix', async () => {
     },
   })
 
-  await axiosInstance.get('https://jsonplaceholder.typicode.com/todos/1')
+  await axiosInstance.get('https://reqres.in/api/users/7')
 
   const [recording] = fs.readdirSync('./recordings')
   expect(recording).toMatchInlineSnapshot(
-    `"todos-1_3439973d2c1ff6cc118d7af4cf797551.json"`,
+    `"api-users-7_1ab499d565acf87f67ac738631d42d07.json"`,
   )
 })
 
 test('returns a cached response when a recording exists', async () => {
   const axiosInstance = axios.create({
-    baseURL: 'https://jsonplaceholder.typicode.com',
+    baseURL: 'https://reqres.in',
   })
 
   const spy = jest.spyOn(axiosInstance.defaults, 'adapter')
 
   useAxiosRecordReplayAdapter({axiosInstance})
-  await axiosInstance.get('/todos/1')
-  await axiosInstance.get('/todos/1')
-  await axiosInstance.get('/todos/1')
+  await axiosInstance.get('/api/users/7')
+  await axiosInstance.get('/api/users/7')
+  await axiosInstance.get('/api/users/7')
 
   expect(spy).toHaveBeenCalledTimes(1)
 })
@@ -119,7 +120,7 @@ test('records a custom request', async () => {
     },
   })
 
-  await axiosInstance.get('https://jsonplaceholder.typicode.com/todos/1')
+  await axiosInstance.get('https://reqres.in/api/users/7')
   const [recording] = fs.readdirSync('./recordings')
   const {request} = JSON.parse(
     fs.readFileSync(`./recordings/${recording}`, 'utf8'),
@@ -127,9 +128,9 @@ test('records a custom request', async () => {
 
   expect(format(request)).toMatchInlineSnapshot(`
     "{
-      'path': '/todos/1'
+      'path': '/api/users/7'
     }"
-  `)
+`)
 })
 
 test('create a custom reponse', async () => {
@@ -142,7 +143,7 @@ test('create a custom reponse', async () => {
     },
   })
 
-  await axiosInstance.get('https://jsonplaceholder.typicode.com/todos/1')
+  await axiosInstance.get('https://reqres.in/api/users/7')
   const [recording] = fs.readdirSync('./recordings')
   const {response} = JSON.parse(
     fs.readFileSync(`./recordings/${recording}`, 'utf8'),